// src/app/core/reports/models/employer-change.model.ts
export interface IEmployerChange {
    is_dupe: boolean;
    linkedin_id: string;
    endMoment: Date;
    startMoment: Date;
    ceo: number;
    founder: number;
    current_employer: string;
    prior_employer: string;
    is_active: number;
    employer_linkedin_id: string;
    employer_history_name: string;
    start_date: Date;
    end_date: Date;
    currently_active: number;
    capsule_id: string;
    data_linkedin_id: string;
    company_data_name: string;
    occupation: string;
    current_occupation: string;
    last_updated: Date;
    educationScore: any;
    is_european: number;
    founderScore: any;
    careerScore: any;
    role_id: number;
    level_id: number;
    firstName: string;
    lastName: string;
    linkedin_company_name: string;
    linkedin_company_id: string;
    industry: string;
    ignored: number;
    location: string;
    publicIdentifier: string;
    current_employer_id: string;
    current_employer_name: string;
    current_start_date: Date;
    industry_match: number;
    contractor_match: number;
    prior_employer_id: string;
    prior_employer_cp_id: string;
    prior_employer_name: string;
    prior_occupation: string;
    ratio: string;
    company_watchlist: number;
    since_string: string;
    end_string: string;
    last_updated_string: string;
}