// src/app/core/reports/models/new-founders.model.ts
export interface INewFoundersEntity {
    linkedin_id: string;
    publicIdentifier: string;
    profile_id: string;
    firstName: string;
    lastName: string;
    current_employer_linkedin_name: string;
    location: string;
    occupation: string;
    last_updated: Date;
    educationScore: number;
    careerScore: number;
    founderScore: number;
    isOnWatchlist: number;
    role_id: number;
    level_id: number;
    is_european: Boolean;
    start_date: Date;
}